+++ src/Zicht/Service/Common/Cache/ArrayMatcher.php
@@ -29,6 +29,11 @@ class ArrayMatcher implements RequestMatcher
      *                 (
      *                     [five] => 5
      *                 )
+     *             [grace_default] => 10
+     *             [grace_attributes] => Array
+     *                 (
+     *                     [five] => 5
+     *                 )
      *         )
      * )
      *
@@ -38,6 +43,14 @@ class ArrayMatcher implements RequestMatcher
     public function __construct(array $config)
     {
         foreach ($config as $method => $properties) {
+            // Ensure grace_default exists for backwards compatability
+            if (!isset($properties['grace_default'])) {
+                $properties['grace_default'] = 0;
+            }
+            // Ensure grace_attributes exists for backwards compatability
+            if (!isset($properties['grace_attributes'])) {
+                $properties['grace_attributes'] = [];
+            }
             $this->config[strtolower($method)] = $properties;
         }
     }
@@ -98,5 +111,24 @@ class ArrayMatcher implements RequestMatcher
             }
         }
         return min($ttls);
+    }
+
+    /**
+     * Return the grace time (in seconds) for the specified request
+     *
+     * @param RequestInterface $request
+     * @return int
+     */
+    public function getGrace(RequestInterface $request)
+    {
+        $config = $this->config[strtolower($request->getMethod())];
+
+        $graces = [$config['grace_default']];
+        foreach ($config['grace_attributes'] as $attribute => $grace) {
+            if ($request->hasAttribute($attribute)) {
+                $graces [] = $grace;
+            }
+        }
+        return min($graces);
     }
 }