+++ src/Zicht/Service/Common/Observers/LoggableException.php
@@ -9,6 +9,7 @@ namespace Zicht\Service\Common\Observers;
  * If a domain exception is loggable, it can tell the Logger observer to use a specific log level for logging the fault.
  *
  * It conveniently extends the Loggable interface to inherit log levels.
+ * @deprecated Remove in next major version
  */
 interface LoggableException
 {