+++ src/Zicht/Service/Common/Observers/RedisLockingCacheTerminateException.php
@@ -1,0 +1,10 @@
+<?php
+/**
+ * @copyright Zicht Online <http://zicht.nl>
+ */
+
+namespace Zicht\Service\Common\Observers;
+
+class RedisLockingCacheTerminateException extends \Exception
+{
+}