+++ src/Zicht/Service/Common/ServiceCall.php
@@ -11,14 +11,20 @@ namespace Zicht\Service\Common;
  */
 class ServiceCall implements ServiceCallInterface
 {
-    /** @var ServiceWrapper|null */
-    private $service = null;
+    /** @var ServiceWrapper */
+    private $service;
 
-    /** @var RequestInterface|null */
-    private $request = null;
+    /** @var RequestInterface */
+    private $request;
 
-    /** @var ResponseInterface|null */
-    private $response = null;
+    /** @var ResponseInterface */
+    private $response;
+
+    /** @var ServiceCallInterface|null */
+    private $parent;
+
+    /** @var bool */
+    private $terminating = false;
 
     /** @var array */
     private $cancelled = [];
@@ -31,13 +37,15 @@ class ServiceCall implements ServiceCallInterface
      * @param RequestInterface $request
      * @param ResponseInterface $response
      * @param ServiceCallInterface|null $parent
+     * @param bool $terminating
      */
-    public function __construct(ServiceWrapper $service, RequestInterface $request, ResponseInterface $response, $parent = null)
+    public function __construct(ServiceWrapper $service, RequestInterface $request, ResponseInterface $response, $parent = null, $terminating = false)
     {
         $this->service = $service;
         $this->request = $request;
         $this->response = $response;
         $this->parent = $parent;
+        $this->terminating = $terminating;
     }
 
     /**
@@ -67,7 +75,10 @@ class ServiceCall implements ServiceCallInterface
     }
 
     /**
-     * Cancel the call, i.e. don't do the request. Only has effect in the notifyBefore() call.
+     * Cancel the call
+     *
+     * The service request is not performed and Observers may alter their behavior by checking $call->isCancelled().
+     * This only has effect in the notifyBefore() call.
      *
      * @param mixed $by
      * @return void
@@ -78,7 +89,7 @@ class ServiceCall implements ServiceCallInterface
     }
 
     /**
-     * Checks if the event was cancelled.
+     * Check if the event was cancelled
      *
      * @param null $by
      * @return bool
@@ -92,13 +103,23 @@ class ServiceCall implements ServiceCallInterface
     }
 
     /**
-     * Returns an array of what classes were responsible for cancelling the request
+     * Returns an array of classes responsible for cancelling the event.
      *
      * @return array
      */
     public function getCancelledBy()
     {
         return $this->cancelled;
+    }
+
+    /**
+     * Returns true when the call is made in terminate mode.
+     *
+     * @return bool
+     */
+    public function isTerminating()
+    {
+        return $this->terminating;
     }
 
     /**
@@ -118,7 +139,7 @@ class ServiceCall implements ServiceCallInterface
     }
 
     /**
-     * @return null|\Zicht\Service\Common\ServiceWrapper
+     * @return ServiceWrapper
      */
     public function getService()
     {