+++ src/Zicht/Service/Common/Soap/Observer/VerboseLoggingObserver.php
@@ -10,7 +10,7 @@ use Zicht\Service\Common\ServiceCallInterface;
 
 /**
  * Adds SOAP request logging to the service call
- *
+ * @deprecated Remove in next major version
  */
 class VerboseLoggingObserver extends ServiceObserverAdapter
 {